Took maintainership of linux-lts-tresor in AUR.
I've redone the PKGBUILD from the current linux-lts PKGBUILD...

This uses linux-lts 3.10.13-1 from Arch's package repository, and the 3.10.13
Linux kernel version, but the latest TRESOR patch is currently for kernel
version 3.8.2.  Caution is strongly advised before using this kernel as it may
be unstable as a result!

I hope TRESOR makes it into the mainline kernel someday.